Create, Delete implemented for Student

+++ src/main/java/com/loginservices/repository/StudentRepository.java
@@ -8,7 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.loginservices.model.Student;
 
 @Repository
-public interface StudentRepository extends CrudRepository<Student, Long> {
+public interface StudentRepository extends CrudRepository<Student, Integer> {
 
 	List<Student> findByEmail(String email);
+	
+	//Student save(Student student);
+	
 }
